use lazy_static http server, restore unauthorized_coverage
remove silo_id from params, it was not used
