+++ nexus/tests/integration_tests/unauthorized.rs
@@ -9,7 +9,6 @@ use super::endpoints::*;
 use crate::integration_tests::silos::SAML_IDP_DESCRIPTOR;
 use dropshot::test_util::ClientTestContext;
 use dropshot::HttpErrorResponseBody;
-use expectorate::assert_contents;
 use headers::authorization::Credentials;
 use http::method::Method;
 use http::StatusCode;
@@ -25,17 +24,7 @@ use nexus_test_utils_macros::nexus_test;
 use omicron_common::api::external::IdentityMetadataCreateParams;
 use omicron_nexus::authn::external::spoof;
 use omicron_nexus::db::fixed_data::silo::DEFAULT_SILO;
-use omicron_nexus::db::identity::Resource;
 use omicron_nexus::external_api::params;
-use openapiv3::OpenAPI;
-use std::collections::BTreeMap;
-
-#[derive(Clone, Debug, Eq, Ord, PartialEq, PartialOrd)]
-struct Operation<'a> {
-    method: &'a str,
-    path: &'a str,
-    label: String,
-}
 
 // This test hits a list Nexus API endpoints using both unauthenticated and
 // unauthorized requests to make sure we get the expected behavior (generally:
@@ -70,28 +59,7 @@ async fn test_unauthorized(cptestctx: &ControlPlaneTestContext) {
     let client = &cptestctx.external_client;
     let log = &cptestctx.logctx.log;
 
-    // Run a httptest server
-    let server = ServerBuilder::new().run().unwrap();
-
-    // Fake some data
-    server.expect(
-        Expectation::matching(request::method_path("HEAD", "/image.raw"))
-            .times(1..)
-            .respond_with(
-                status_code(200).append_header(
-                    "Content-Length",
-                    format!("{}", 4096 * 1000),
-                ),
-            ),
-    );
-
-    server.expect(
-        Expectation::matching(request::method_path("GET", "/descriptor"))
-            .times(1..)
-            .respond_with(status_code(200).body(SAML_IDP_DESCRIPTOR)),
-    );
-
-    // Create test data with hardcoded values.
+    // Create test data.
     info!(log, "setting up resource hierarchy");
     for request in &*SETUP_REQUESTS {
         NexusRequest::objects_post(client, request.url, &request.body)
@@ -101,121 +69,12 @@ async fn test_unauthorized(cptestctx: &ControlPlaneTestContext) {
             .unwrap();
     }
 
-    // Create test data with non-hardcoded values.
-
-    // Create a GlobalImage
-    let demo_image_create: params::ImageCreate = params::ImageCreate {
-        identity: IdentityMetadataCreateParams {
-            name: DEMO_IMAGE_NAME.clone(),
-            description: String::from(""),
-        },
-        source: params::ImageSource::Url(server.url("/image.raw").to_string()),
-        block_size: params::BlockSize::try_from(4096).unwrap(),
-    };
-
-    NexusRequest::objects_post(client, "/images", &demo_image_create)
-        .authn_as(AuthnMode::PrivilegedUser)
-        .execute()
-        .await
-        .unwrap();
-
-    // Create a SAML identity provider
-    let demo_silo_saml_idp_create = params::SiloSamlIdentityProviderCreate {
-        identity: IdentityMetadataCreateParams {
-            name: "demo-saml-provider".to_string().parse().unwrap(),
-            description: "a demo provider".to_string(),
-        },
-
-        silo_id: DEFAULT_SILO.id().clone(),
-        idp_metadata_url: server.url("/descriptor").to_string(),
-
-        idp_entity_id: "entity_id".to_string(),
-        sp_client_id: "client_id".to_string(),
-        acs_url: "http://acs".to_string(),
-        slo_url: "http://slo".to_string(),
-        technical_contact_email: "technical@fake".to_string(),
-
-        public_cert: None,
-        private_key: None,
-    };
-
-    NexusRequest::objects_post(
-        client,
-        "/silos/default-silo/saml_identity_provider",
-        &demo_silo_saml_idp_create,
-    )
-    .authn_as(AuthnMode::PrivilegedUser)
-    .execute()
-    .await
-    .unwrap();
-
-    let mut non_hardcoded_endpoints = vec![
-        /* Project images */
-        VerifyEndpoint {
-            url: &*DEMO_PROJECT_URL_IMAGES,
-            visibility: Visibility::Protected,
-            allowed_methods: vec![
-                AllowedMethod::GetUnimplemented,
-                AllowedMethod::Post(
-                    serde_json::to_value(&demo_image_create).unwrap(),
-                ),
-            ],
-        },
-        VerifyEndpoint {
-            url: &*DEMO_PROJECT_IMAGE_URL,
-            visibility: Visibility::Protected,
-            allowed_methods: vec![
-                AllowedMethod::GetUnimplemented,
-                AllowedMethod::Delete,
-            ],
-        },
-        /* Global images */
-        VerifyEndpoint {
-            url: "/images",
-            visibility: Visibility::Public,
-            allowed_methods: vec![
-                AllowedMethod::Get,
-                AllowedMethod::Post(
-                    serde_json::to_value(&demo_image_create).unwrap(),
-                ),
-            ],
-        },
-        VerifyEndpoint {
-            url: &*DEMO_GLOBAL_IMAGE_URL,
-            visibility: Visibility::Protected,
-            allowed_methods: vec![AllowedMethod::Get, AllowedMethod::Delete],
-        },
-        /* Silo identity providers */
-        VerifyEndpoint {
-            url: "/silos/default-silo/saml_identity_provider",
-            visibility: Visibility::Public, // Users can see their own silo! This includes USER_TEST_UNPRIVILEGED
-            allowed_methods: vec![AllowedMethod::Post(
-                serde_json::to_value(&demo_silo_saml_idp_create).unwrap(),
-            )],
-        },
-        VerifyEndpoint {
-            url:
-                "/silos/default-silo/saml_identity_provider/demo-saml-provider",
-            visibility: Visibility::Protected,
-            allowed_methods: vec![AllowedMethod::Get],
-        },
-    ];
-
-    let mut endpoints: Vec<VerifyEndpoint> = Vec::with_capacity(
-        VERIFY_ENDPOINTS.len() + non_hardcoded_endpoints.len(),
-    );
-    endpoints.extend(VERIFY_ENDPOINTS.clone());
-    endpoints.append(&mut non_hardcoded_endpoints);
-
     // Verify the endpoints.
     info!(log, "verifying endpoints");
     print!("{}", VERIFY_HEADER);
-    for endpoint in &endpoints {
+    for endpoint in &*VERIFY_ENDPOINTS {
         verify_endpoint(&log, client, endpoint).await;
     }
-
-    // Validate unauthorized coverage
-    unauthorized_coverage(&endpoints);
 }
 
 const VERIFY_HEADER: &str = r#"
@@ -268,6 +127,31 @@ struct SetupReq {
 }
 
 lazy_static! {
+    pub static ref HTTP_SERVER: httptest::Server = {
+        // Run a httptest server
+        let server = ServerBuilder::new().run().unwrap();
+
+        // Fake some data
+        server.expect(
+            Expectation::matching(request::method_path("HEAD", "/image.raw"))
+                .times(1..)
+                .respond_with(
+                    status_code(200).append_header(
+                        "Content-Length",
+                        format!("{}", 4096 * 1000),
+                    ),
+                ),
+        );
+
+        server.expect(
+            Expectation::matching(request::method_path("GET", "/descriptor"))
+                .times(1..)
+                .respond_with(status_code(200).body(SAML_IDP_DESCRIPTOR)),
+        );
+
+        server
+    };
+
     /// List of requests to execute at setup time
     static ref SETUP_REQUESTS: Vec<SetupReq> = vec![
         // Create a separate Silo (not used for anything else)
@@ -314,6 +198,16 @@ lazy_static! {
         SetupReq {
             url: &*DEMO_PROJECT_URL_INSTANCES,
             body: serde_json::to_value(&*DEMO_INSTANCE_CREATE).unwrap(),
+        },
+        // Create a GlobalImage
+        SetupReq {
+            url: "/images",
+            body: serde_json::to_value(&*DEMO_IMAGE_CREATE).unwrap(),
+        },
+        // Create a SAML identity provider
+        SetupReq {
+            url: "/silos/default-silo/saml_identity_provider",
+            body: serde_json::to_value(&*SILO_SAML_IDENTITY_PROVIDER).unwrap(),
         },
     ];
 }
@@ -655,130 +549,4 @@ fn record_operation(whichtest: WhichTest<'_>) {
     } else {
         print!("{}", c);
     }
-}
-
-fn unauthorized_coverage(endpoints: &Vec<VerifyEndpoint>) {
-    // Load the OpenAPI schema for Nexus's public API.
-    let schema_path = "../openapi/nexus.json";
-    let schema_contents = std::fs::read_to_string(&schema_path)
-        .expect("failed to read Nexus OpenAPI spec");
-    let spec: OpenAPI = serde_json::from_str(&schema_contents)
-        .expect("Nexus OpenAPI spec was not valid OpenAPI");
-
-    // Take each operation that we find in the OpenAPI spec, make a regular
-    // expression that we can use to match against it (more on this below), and
-    // throw them into a BTreeMap.
-    let mut spec_operations: BTreeMap<Operation, regex::Regex> = spec
-        .operations()
-        .map(|(path, method, op)| {
-            // We're going to take URLs from our test cases and match them
-            // against operations in the API spec.  The URLs from the API spec
-            // contain variables (e.g., "/instances/{instance_name}").  Our test
-            // cases have those variables filled in already (e.g.,
-            // "/instances/my-instance").
-            //
-            // To match a URL from the test case against one from the OpenAPI
-            // spec, we're going to:
-            //
-            // - use a regular expression to replace `{varname}` in the API
-            //   spec's URL with `[^/]+` (one or more non-slash characters)
-            //
-            // - use that string as the basis for a second regex that we'll
-            //   store with the Operation.  We'll use this second regex to match
-            //   URLs to their operation.
-            //
-            // This is slow (lookups will take time linear in the total number
-            // of API endpoints) and a little cheesy, but it's expedient and
-            // robust enough for our purposes.
-            //
-            // This will fail badly if it turns out that the URL contains any
-            // characters that would be interpreted specially by the regular
-            // expression engine.  So let's check up front that those aren't
-            // present.
-            assert!(
-                path.chars().all(|c| c.is_ascii_alphanumeric()
-                    || c == '_'
-                    || c == '-'
-                    || c == '{'
-                    || c == '}'
-                    || c == '/'),
-                "unexpected character in URL: {:?}",
-                path
-            );
-            let re = regex::Regex::new("/\\{[^}]+\\}").unwrap();
-            let regex_path = re.replace_all(path, "/[^/]+");
-            let regex = regex::Regex::new(&format!("^{}$", regex_path))
-                .expect("modified URL string was not a valid regex");
-            let label = op
-                .operation_id
-                .clone()
-                .unwrap_or(String::from("unknown operation-id"));
-            (Operation { method, path, label }, regex)
-        })
-        .collect();
-
-    // Go through each of the authz test cases and match each one against an
-    // OpenAPI operation.
-    let mut unexpected_endpoints = String::from(
-        "API endpoints tested by unauthorized.rs but not found \
-        in the OpenAPI spec:\n",
-    );
-    for v in endpoints {
-        for m in &v.allowed_methods {
-            let method_string = m.http_method().to_string().to_uppercase();
-            let found = spec_operations.iter().find(|(op, regex)| {
-                op.method.to_uppercase() == method_string
-                    && regex.is_match(v.url)
-            });
-            if let Some((op, _)) = found {
-                println!(
-                    "covered: {:40} ({:6?} {:?}) (by {:?})",
-                    op.label, op.method, op.path, v.url
-                );
-                let op = op.clone();
-                spec_operations.remove(&op);
-            } else {
-                unexpected_endpoints
-                    .push_str(&format!("{:6} {:?}", method_string, v.url));
-            }
-        }
-    }
-
-    println!("-----");
-
-    // If you're here because this assertion failed, we found an endpoint tested
-    // by "unauthorized.rs" that's not in the OpenAPI spec.  This could happen
-    // if you're adding a test for an endpoint that's marked "unpublished".  In
-    // that case, you might just allow expectorate to add this endpoint to the
-    // allowlist here.
-    assert_contents(
-        "tests/output/unexpected-authz-endpoints.txt",
-        &unexpected_endpoints,
-    );
-
-    // Check for uncovered endpoints (endpoints that are in the OpenAPI spec but
-    // not tested by the authz tests).
-    let mut uncovered_endpoints =
-        "API endpoints with no coverage in authz tests:\n".to_string();
-    for op in spec_operations.keys() {
-        uncovered_endpoints.push_str(&format!(
-            "{:40} ({:6} {:?})\n",
-            op.label, op.method, op.path
-        ));
-    }
-
-    // If you're here because this assertion failed, check that if you've added
-    // any API operations to Nexus, you've also added a corresponding test in
-    // "unauthorized.rs" so that it will automatically be checked for its
-    // behavior for unauthenticated and unauthorized users.  DO NOT SKIP THIS.
-    // Even if you're just adding a stub, see [`Nexus::unimplemented_todo()`].
-    // If you _added_ a test that covered an endpoint from the allowlist --
-    // hooray!  Just delete the corresponding line from this file.  (Why is this
-    // not `expectorage::assert_contents`?  Because we only expect this file to
-    // ever shrink, which is easy enough to fix by hand, and we don't want to
-    // make it easy to accidentally add things to the allowlist.)
-    let expected_uncovered_endpoints =
-        std::fs::read_to_string("tests/output/uncovered-authz-endpoints.txt")
-            .expect("failed to load file of allowed uncovered endpoints");
-    assert_eq!(expected_uncovered_endpoints, uncovered_endpoints);
 }
