+++ openapi/nexus.json
@@ -7490,10 +7490,6 @@
             "description": "optional request signing public certificate (base64 encoded der file)",
             "type": "string"
           },
-          "silo_id": {
-            "type": "string",
-            "format": "uuid"
-          },
           "slo_url": {
             "description": "service provider endpoint where the idp should send log out requests",
             "type": "string"
@@ -7513,7 +7509,6 @@
           "idp_entity_id",
           "idp_metadata_url",
           "name",
-          "silo_id",
           "slo_url",
           "sp_client_id",
           "technical_contact_email"
